backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.10.0
python-multipart==0.0.6
pydub==0.25.1
mutagen==1.46.0
rarfile==4.1
httpx==0.27.2
PyJWT==2.8.0
zipstream-ng==1.7.4
